exp_disponibilidad/replica-set/requirements.txt
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.15
Werkzeug==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary
itsdangerous
flask
flask_cors
aiven-client
python-dotenv
eventlet
pika
flask_coney
redis




